Added solution - LeetHub

+++ Make array elements unique - GFG/make-array-elements-unique.java
@@ -40,7 +40,7 @@ class Solution {
         // Code here
       Arrays.sort(arr);
       
-      long ans =0;
+      int ans =0;
       
       for(int i=0; i<N-1; i++){
             if(arr[i]>=arr[i+1]){
